AIStudioEnterprise/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.1
google-auth==2.30.0
jwt==1.3.1
mysqlclient==2.2.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.0.3
python-decouple==3.8
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.1
